Refactor: Extract useSuperglue tests from hooks.spec.jsx
Closes #207
